day16: part 1, Ensure numbers are large enough to hold values
Can't unpack 40 bits into 32 bits!

+++ day16/src/main.rs
@@ -5,7 +5,7 @@ use std::io::ErrorKind;
 
 use log;
 
-type Num = u32;
+type Num = u64;
 
 mod bits;
 use bits::*;
